Began work on storing all global styles in a <style> tag within the <head> of the document. This change has allowed for the following:
1. Less element style updates to the mousetip other than the top/left
   styles used to follow the cursor.
2. A new animation flag (true by default) that when true adds additional
   functionality to fade/slide the mousrtip in/out upon
   creation/deletion respectively.

With this comes a change to the position global setting, which now must
be an array as opposed to a string. Global settings will see an overhaul
in the near future, so this may change (including its name or how to
declare it).

The readme.md also has been updated to include the position global
setting change to keep track of the changes later on.