Unique identifiers
When initializing multiple instances of MouseTip, it includes the styles
only once; This creates a problem for global style/animation overrides,
as the existing code creates these overrides for the class, which is
shared between all instances, causing the overrides to affect all
instances. To combat this, a unique identifier is now generated and the
overrides target this ID instead of the class. This ID can also be
manually set within the constructor object.