Default style changes
In previous updates, different background/foreground colors were used
for the mousetip when in light/dark mode; Not only is this opinionated
styling, but it didn't look as good on light mode.

Instead, the different background/foreground colors were dropped in
favor of one set used across light/dark modes. The background color was
lightened a bit to stand out better in dark mode.